split recipe page into smaller components
